+++ contracts/contracts/gateway/GatewayGetterFacet.sol
@@ -139,6 +139,10 @@ contract GatewayGetterFacet {
         return (s.postbox[id]);
     }
 
+    function postboxMsgs() external view returns (bytes32[] memory) {
+        return (s.postboxKeys.values());
+    }
+
     /// @notice Returns the majority percentage required for certain consensus or decision-making processes.
     function majorityPercentage() external view returns (uint64) {
         return s.majorityPercentage;