+++ ext/frc42_dispatch/macros/tests/naming/non-capital-start.rs
@@ -1,0 +1,13 @@
+// Copyright 2022-2024 Protocol Labs
+// SPDX-License-Identifier: MIT
+//
+// Forked from https://github.com/filecoin-project/actors-utils with assumed MIT license
+// as per Cargo.toml: https://github.com/filecoin-project/actors-utils/blob/7628cd8d39dafcc6035f28e350cdb0cccbea5ab4/frc42_dispatch/Cargo.toml#L5
+//
+// License headers added post-fork.
+use frc42_macros::method_hash;
+
+fn main() {
+	// should panic because the name starts with non-capital letter
+	let _str_hash = method_hash!("noPlaceForCamelCase");
+}